requirements.txt
fastapi==0.115.12
python-dotenv==1.1.0
groq==0.22.0
sentence-transformers==4.1.0
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.6.1
